Move process_instruction definitions to runtime
