+++ runtime/src/lib.rs
@@ -22,6 +22,7 @@ pub mod log_collector;
 pub mod message_processor;
 mod native_loader;
 pub mod nonce_utils;
+pub mod process_instruction;
 pub mod rent_collector;
 pub mod serde_snapshot;
 pub mod snapshot_package;
